requirements.txt
aioredis==1.2.0
asgiref==2.3.2
asn1crypto==0.24.0
astroid==2.0.4
async-timeout==3.0.1
attrs==18.2.0
autobahn==18.10.1
Automat==0.7.0
automium==0.2.6
automium-web==0.1.1
certifi==2018.10.15
cffi==1.12.3
channels==2.1.5
channels-redis==2.3.1
chardet==3.0.4
Click==7.0
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.7
daphne==2.2.2
defusedxml==0.5.0
Django==2.2
django-allauth==0.38.0
django-cors-headers==2.4.0
django-rest-auth==0.9.3
django-rest-framework==0.1.0
django-rest-swagger==2.2.0
djangorestframework==3.9.0
djangorestframework-jwt==1.11.0
Flask==1.1.1
gunicorn==19.7.1
hiredis==0.2.0
hyperlink==18.0.0
idna==2.7
incremental==17.5.0
isort==4.3.4
itsdangerous==1.1.0
itypes==1.1.0
Jinja2==2.10.1
lazy-object-proxy==1.3.1
legacy==0.1.6
MarkupSafe==1.1.1
mccabe==0.6.1
msgpack==0.5.6
oauthlib==2.1.0
openapi-codec==1.3.2
psycopg2==2.8.3
psycopg2-binary==2.7.5
pyasn1==0.4.7
pycparser==2.19
PyHamcrest==1.9.0
PyJWT==1.7.1
pylint==2.1.1
python-dotenv==0.10.3
python3-openid==3.1.0
pytz==2018.7
redis==3.3.8
requests==2.20.1
requests-oauthlib==1.0.0
simplejson==3.16.0
six==1.11.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
sqlparse==0.3.0
Twisted==18.9.0
txaio==18.8.1
typed-ast==1.1.0
uritemplate==3.0.0
urllib3==1.24.3
Werkzeug==0.15.6
whitenoise==3.3.1
wrapt==1.10.11
zope.interface==4.6.0
